requirements.txt
Flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==4.5.0

#pip install -r requirements.txt
